requirements.txt
flask
flask-login
mysql-connector-python
werkzeug
python-dotenv
flask-mail